+++ src/store.h
@@ -35,14 +35,14 @@ public:
     virtual const std::string& sexp_type() const = 0;// the return type will be more complex when we deal with richer queries
     virtual const std::string& store_kind() const { return store_k;};// typed, generic, specialized...
 
-    virtual SEXP get_value(size_t index) = 0;
+    virtual SEXP get_value(uint64_t index) = 0;
 
-    virtual const sexp_hash& get_hash(size_t index) const = 0;
+    virtual const sexp_hash& get_hash(uint64_t index) const = 0;
 
     virtual size_t nb_values() const = 0;
 
     virtual SEXP get_metadata(SEXP val) const = 0;
-    virtual SEXP get_metadata(size_t index) const = 0;
+    virtual SEXP get_metadata(uint64_t index) const = 0;
 
     virtual const fs::path& description_path() const = 0;
